LPS-184314 Extract methods from CompanyIndexFactory to CompanyIndexFactoryHelper, apply CompanyIndexFactoryHelper to usages, apply IndexFactoryCompanyIdRegistryUtil to left usages and change UnprocessedExceptionCheck sf suppression file
